Adds dependency for creating a workmanager and uploadEventWorkManager
