+++ data/build.gradle.kts
@@ -88,6 +88,8 @@ dependencies {
     implementation("androidx.security:security-crypto-ktx:1.1.0-alpha05")
     implementation("androidx.work:work-runtime-ktx:2.8.0")
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:1.1.5")
+    implementation("androidx.work:work-runtime-ktx:2.8.0")
+    implementation("androidx.hilt:hilt-work:1.0.0")
 
     testImplementation(tests.hilt.android.testing)
     kaptTest(tests.hilt.compiler)
