+++ domain/src/main/java/me/androidbox/domain/work_manager/UploadEventWorker.kt
@@ -1,0 +1,7 @@
+package me.androidbox.domain.work_manager
+
+import me.androidbox.domain.authentication.model.Event
+
+interface UploadEventWorker {
+    suspend fun uploadEvent(event: Event, isEditModel: Boolean)
+}
