+++ presentation/src/main/java/me/androidbox/presentation/HomeActivity.kt
@@ -4,21 +4,15 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.runtime.*
-import androidx.compose.runtime.livedata.observeAsState
-import androidx.core.net.toUri
+import androidx.compose.runtime.collectAsState
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.navigation.compose.rememberNavController
-import androidx.work.*
 import dagger.hilt.android.AndroidEntryPoint
 import me.androidbox.domain.authentication.ResponseState
-import me.androidbox.presentation.event.viewmodel.EventViewModel
 import me.androidbox.presentation.login.viewmodel.HomeViewModel
 import me.androidbox.presentation.navigation.NavigationGraph
 import me.androidbox.presentation.navigation.Screen
 import me.androidbox.presentation.ui.theme.BusbyTaskyTheme
-import me.androidbox.presentation.worker.UploadEventWorker
-import java.util.*
 
 @AndroidEntryPoint
 class HomeActivity : ComponentActivity() {
@@ -32,10 +26,6 @@ class HomeActivity : ComponentActivity() {
                 homeViewModel.authenticationState.value == null
             }
         }
-
-        /* TODO Just for testing inserting and fetching */
-        val eventViewModel by viewModels<EventViewModel>()
-        eventViewModel.insertEvent()
 
         setContent {
             val authenticatedState = homeViewModel.authenticationState.collectAsState()
